Rename the directory from default to defaultservicename to conform to go style guide. Wire admission controller into the apiserver
