+++ cmd/apiserver/app/plugins.go
@@ -22,5 +22,5 @@ package app
 import (
 	// Admission policies
 	_ "github.com/kubernetes-incubator/service-catalog/plugin/pkg/admission/namespace/lifecycle"
-	_ "github.com/kubernetes-incubator/service-catalog/plugin/pkg/admission/serviceplan/default"
+	_ "github.com/kubernetes-incubator/service-catalog/plugin/pkg/admission/serviceplan/defaultserviceplan"
 )
